requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langgraph
langchain-google-genai
langchain-groq
langchain-openai
duckduckgo-search
supabase
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
sse-starlette
pytest
pytest-asyncio
